% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cap_ages.R
\name{Cap_ages}
\alias{Cap_ages}
\title{Title}
\usage{
Cap_ages(data, min_age = -1, max_age = Inf)
}
\arguments{
\item{data}{a data frame containing the columns modal_age_closest, modal_age_low, age, corresponding to the modal age closest to the mean, the lowest modal age, and an age-reading, respectively. Typically the output generated by the function Compute_modes.}

\item{min_age}{The minimum age, defaults to -1}

\item{max_age}{The  maximum age, defaults to infinity}
}
\value{
a data frame with all ages capped at min_age and max_age
}
\description{
Title
}
\examples{
file_path = system.file("extdata",
                        "NSS_herring_scales.csv",
                         package = "ParametricAEM")
data_nss_scales = read.csv(file_path, sep = ",")
data_exp = data_nss_scales[data_nss_scales$expertise == 1, ] # Extract experts
data_exp = data_exp[data_exp$age > 0, ]
data_analysis = Compute_modes(data_exp)
data_analysis_capped = Cap_ages(data_analysis)
}
