% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_modes.R
\name{Compute_modes}
\alias{Compute_modes}
\title{Compute modal ages per fish}
\usage{
Compute_modes(data)
}
\arguments{
\item{data}{A data frame with one row per age-reading containing at least the columns:
FishID, a unique identifier per fish
age, an age-reading}
}
\value{
A data frame with three added columns:
modal_age_low, the lowest modal age
modal_age_closest, the modal age closest to the mean
weight, the corresponding weights for the modes (differs from 1 when the modal age closest to the mean is not unique)
}
\description{
Compute modal ages per fish
}
\examples{
file_path = system.file("extdata",
                        "NSS_herring_scales.csv",
                         package = "ParametricAEM")
data_nss_scales = read.csv(file_path, sep = ",")
data_exp = data_nss_scales[data_nss_scales$expertise == 1, ] # Extract experts
data_exp = data_exp[data_exp$age > 0, ]
data_analysis = Compute_modes(data_exp)
}
