% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_model.R
\name{Estimate_model}
\alias{Estimate_model}
\title{Estimate the parametric age-reading error matrix}
\usage{
Estimate_model(
  data,
  min_age,
  max_age,
  start = c(2, 0, 0.5, 0, 0.2),
  estimate = rep(1, 5)
)
}
\arguments{
\item{data}{a data frame containing the columns modal_age_closest, modal_age_low, age, corresponding to the modal age closest to the mean, the lowest modal age, and an age-reading, respectively. No ages are lower than min_age, and no ages are higher than max_age. Typically the output generated by the function Cap_ages.}

\item{min_age}{Integer, the minimum age}

\item{max_age}{Integer, the maximum age}

\item{start}{Numerical vector with length 5. Starting values for the parameters in the order beta0, beta1, alpha0, alpha1, phi.}

\item{estimate}{Vector of length 5 with indicators for whether each parameter should be estimated (1) or fixed (0). Defaults to estimating all parameters.}
}
\value{
a list with par_est: the estimated parameters and corresponding standard deviations, Sigma: the covariance matrix for the estimated parameters, loglik: the log likelihood
}
\description{
Estimate the parametric age-reading error matrix
}
\examples{
file_path = system.file("extdata",
                        "NSS_herring_scales.csv",
                         package = "ParametricAEM")
data_nss_scales = read.csv(file_path, sep = ",")
data_exp = data_nss_scales[data_nss_scales$expertise == 1, ] # Extract experts
data_exp = data_exp[data_exp$age > 0, ]
data_analysis = Compute_modes(data_exp)
data_analysis_capped = Cap_ages(data_analysis)
fit = Estimate_model(data = data_analysis_capped,
                     min_age = 3,
                     max_age = 12,
                     start = c(2, 0, 0, 0, 0.2),
                     estimate = c(1, 1, 0, 0, 1))
}
