% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_model.R
\name{negloglik}
\alias{negloglik}
\title{Negative log likelihood. This function is called by Estimate_model.}
\usage{
negloglik(param, data, par.fixed, min_age, max_age, parnames, lowmode)
}
\arguments{
\item{param}{Numerical vector with values for parameters that are estimated}

\item{data}{a data frame containing the columns modal_age_closest, modal_age_low, age, corresponding to the modal age closest to the mean, the lowest modal age, and an age-reading, respectively. No ages are lower than min_age, and no ages are higher than max_age. Typically the output generated by the function Cap_ages.}

\item{par.fixed}{Numerical vector of values for fixed parameters}

\item{min_age}{Integer, the minimum age}

\item{max_age}{Integer, the maximum age}

\item{parnames}{Vector of all parameter names in the order: estimated parameters, fixed parameters}

\item{lowmode}{Indicator for whether we should use the lowest modal age (1). Default is to use the modal age closest to the mean (0).}
}
\value{
Float, the negative log likelihood
}
\description{
Negative log likelihood. This function is called by Estimate_model.
}
