% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_model.R
\name{loglik}
\alias{loglik}
\title{Title}
\usage{
loglik(theta, min_age, max_age, data, lowmode)
}
\arguments{
\item{theta}{Numerical vector with values for the estimated parameters}

\item{min_age}{The minimum age}

\item{max_age}{The maximum age}

\item{data}{a data frame containing the columns modal_age_closest, modal_age_low, age, corresponding to the modal age closest to the mean, the lowest modal age, and an age-reading, respectively. No ages are lower than min_age, and no ages are higher than max_age. Typically the output generated by the function Cap_ages.}

\item{lowmode}{Indicator for whether we should use the lowest modal age (1).}
}
\value{
Float, the likelihood
}
\description{
Title
}
